+++ app/lib/cognito.server.ts
@@ -91,22 +91,6 @@ export async function getCognitoUserFromSub(sub: string) {
   return user
 }
 
-// Gets user from Cognito by sub to be indexed into OpenSearch
-export async function getUserForOpenSearch(sub: string) {
-  const user = await getCognitoUserFromSub(sub)
-  if (!user) throw new Response(null, { status: 404 })
-  const groups = user.Username
-    ? ((await getUserGroupStrings(user.Username)) ?? [])
-    : []
-  return {
-    sub: extractAttributeRequired(user.Attributes, 'sub'),
-    email: extractAttributeRequired(user.Attributes, 'email'),
-    name: extractAttribute(user.Attributes, 'name'),
-    affiliation: extractAttribute(user.Attributes, 'custom:affiliation'),
-    groups,
-  }
-}
-
 export async function listUsers(
   filterString: string
 ): Promise<Omit<User, 'idp' | 'cognitoUserName' | 'groups'>[]> {
